sphere

+++ volumeofsphere.cpp
@@ -5,7 +5,7 @@ int main(){
     int n;
     cout<<"Enter value of n:\n";
     cin>>n;
-    int sphere=(4*3.14*n*n*n)/3;
+    float sphere=(4*3.14*n*n*n)/3;
     cout<<sphere;
     return 0;
